# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server API.",
        "title": "dm-trade API",
        "termsOfService": "http://example.com/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "description": "Get categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get user orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/domain.Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "post": {
                "description": "create order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "User Registration Data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateOrderInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Get Products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get Products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/domain.Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/products/search": {
            "post": {
                "description": "Search for products based on various filters and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Search Products",
                "parameters": [
                    {
                        "description": "Search and filter parameters",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.GetProductsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/domain.Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.GetProductsQuery": {
            "type": "object",
            "properties": {
                "in_stock": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "max_price": {
                    "type": "integer"
                },
                "min_price": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "skip": {
                    "type": "integer"
                },
                "sorted_type": {
                    "type": "integer"
                }
            }
        },
        "domain.Order": {
            "type": "object",
            "properties": {
                "User": {
                    "$ref": "#/definitions/domain.User"
                },
                "address": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "deliveryType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orderDate": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OrderProduct"
                    }
                },
                "recipientEmail": {
                    "type": "string"
                },
                "recipientName": {
                    "type": "string"
                },
                "recipientPhone": {
                    "type": "string"
                },
                "totalPrice": {
                    "type": "number"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "domain.OrderProduct": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/domain.Order"
                },
                "orderId": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/domain.Product"
                },
                "productId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "totalPrice": {
                    "type": "number"
                }
            }
        },
        "domain.Product": {
            "type": "object",
            "properties": {
                "article": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "domain.User": {
            "type": "object",
            "properties": {
                "blocked": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "service.CreateOrderInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "delivery_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.ProductInput"
                    }
                },
                "recipient_email": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                },
                "recipient_phone": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "service.ProductInput": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_article": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "v1.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}